Add a few missing renames
Rename alfg, getid, strnlen, and timing headers
and code files to be consistent.
