+++ src/mca/ptl/base/ptl_base_connection_hdlr.c
@@ -40,8 +40,8 @@
 #include "src/server/pmix_server_ops.h"
 #include "src/util/pmix_argv.h"
 #include "src/util/pmix_error.h"
-#include "src/util/getid.h"
-#include "src/util/strnlen.h"
+#include "src/util/pmix_getid.h"
+#include "src/util/pmix_strnlen.h"
 
 #include "src/mca/ptl/base/base.h"
 #include "src/mca/ptl/base/ptl_base_handshake.h"
