+++ src/runtime/pmix_params.c
@@ -38,7 +38,7 @@
 #include "src/mca/base/pmix_mca_base_var.h"
 #include "src/runtime/pmix_rte.h"
 #include "src/server/pmix_server_ops.h"
-#include "src/util/timings.h"
+#include "src/util/pmix_timings.h"
 
 #if PMIX_ENABLE_TIMING
 char *pmix_timing_output = NULL;
